refactor #step1 - 클래스 이름을 변경합니다
