--- src/main/java/camp/nextstep/edu/calculator/Validator.java
@@ -1,5 +1,0 @@
-package camp.nextstep.edu.calculator;
-
-public interface Validator<T> {
-    void validate(T string);
-}
