+++ src/test/java/camp/nextstep/edu/calculator/MockExpressionValidator.java
@@ -1,6 +1,6 @@
 package camp.nextstep.edu.calculator;
 
-public class MockExpressionValidator implements Validator<String> {
+public class MockExpressionValidator implements ExpressionValidator {
 
     @Override
     public void validate(String string) {
